+++ ex3/only_mpi.c
@@ -36,7 +36,7 @@ int main(int argc, char *argv[]) {
     double *v = (double *) malloc(n * sizeof(double));
     if (my_rank == 0) {
         for (uint16_t i = 1; i <= n; i++) {
-            v[i-1] = 1 / (double)(i * i);
+            v[i-1] = 1. / (i * i);
         }
     }
     // Broadcast the generated vector to every rank
